FO-2392: Dekomponer henvendelse-inputfelt og style av ny dialog-desktop
